[TASK] Introduce `rector` as local development tool  (#102)
* [TASK] Install rector independently in a subdirectory

This prevents conflicts with typo3/cms-install

* [TASK] Introduce `rector` as local development tool

`rector` is a tool to help with code updates, for php code, but also
application code, which includes TYPO3 by community provided specific
rule-sets.

Due to dependency version conflicts from `rector` and `typo3/cms-install`
for the `nikic parser` dependency it is not possible to have it required
in the root `composer.json`. Until rector releases a phar based binary,
installing rector in a subfolder with a dedicated `composer.json` is a
reasonable and valid workaround.

This change introduce the `rector` tool as local development tool chain,
mainly adding it to the `Build/Scripts/runTests.sh` execution dispatcher.

Added `runTests.sh` commands:

```terminal
./Build/Scripts/runTests.sh -s composerUpdateRector

./Build/Scripts/runTests.sh -s rector
./Build/Scripts/runTests.sh -s rector -n
```

The `-n` option can be used for rector to use the tool in the `dry-run`
mode, not applying changes.

Rector based changes are also applied running rector.

Used command(s):

```terminal
\
  ./Build/Scripts/runTests.sh -s composerUpdateRector ; \
  ./Build/Scripts/runTests.sh -s rector
```

Note: It could be considered to make the rector install locked using a
committed `composer.lock` file and add it as CI code style check in a
dedicated follow-up change.