[TASK] Improve `unit` and `functional` test execution experience
Basic unit and functional testing capabilities has been added
recently. That allows creating unit and functional tests, and
execution of them with manually invoking PHPUnit along with a
corresponding config and in case of functional tests also the
environment variables for the Database to use.

This requires a lot of knowledge and also the steps to provide
empty database systems, and is not really contribution friendly.

The existing `Build/Scripts/runTests.sh` is now modified to ...

* add `-d <sqlite,mariadb,mysql,postgres>` option to define the
  database service to spawn and use when executing functional
  tests.
* add `-a <mysqli|pdo_mysql>` option to select the PHP database
  driver to use for functional tests for `MySQL` and `MariaDB`.
* add `-i` option to specify the database version to use, with
  version constraint and default handling based on the selected
  database vendor with `-d`.
* add `-x` option to enable xdebug capabilities.
* add `-y <port>` to specify the xdebug port to use with `-x`.
* modify the help text to add full description of the added
  options and possible values.
* add helper methods and display database informations at the
  end of script executions.
* add `-s unit` and `-s functional` along with the execution
  code along with the other options.

This also includes a minor bugfix to use the correct docker
hostname for podman users, when using `-x` to start xdebug.